pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["nikita <misharinnikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.8.1"
sqlalchemy = "^1.4.42"
sanic = "^22.9.0"
asyncpg = "^0.26.0"
psycopg2 = "^2.9.4"
pyjwt = "^2.6.0"
pydantic = "^1.10.2"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
sanic-ext = "^22.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
